refactored firstcollisionsnsolver.h and added the header code
