+++ include/solvers/firstcollisionsnsolver.h
